+++ selectrum.el
@@ -921,7 +921,15 @@ into the user input area to start with."
   (interactive)
   (when selectrum--current-candidate-index
     (setq selectrum--current-candidate-index
-          (max (if selectrum--match-required-p 0 -1)
+          (max (if (and selectrum--match-required-p
+                        (cond (minibuffer-completing-file-name
+                               (not (file-exists-p
+                                     (selectrum--current-input))))
+                              (t
+                               (not (string-empty-p
+                                     (selectrum--current-input))))))
+                   0
+                 -1)
                (1- selectrum--current-candidate-index)))))
 
 (defun selectrum-next-candidate ()
@@ -1014,7 +1022,12 @@ Zero means to select the current user input."
                    (min (1- (prefix-numeric-value arg))
                         (1- (length selectrum--refined-candidates)))
                  selectrum--current-candidate-index)))
-    (when (or index (not selectrum--match-required-p))
+    (when (or (not selectrum--match-required-p)
+              (and index (>= index 0))
+              (and minibuffer-completing-file-name
+                   (file-exists-p (selectrum--current-input)))
+              (string-empty-p
+               (selectrum--current-input)))
       (selectrum--exit-with
        (selectrum--get-candidate index)))))