+++ CHANGELOG.md
@@ -169,6 +169,8 @@ The format is based on [Keep a Changelog].
   default completion UI ([#101]).
 * Text properties are now stripped for standard completion functions
   ([#107], [#108]).
+* You can now select and submit empty input and for file prompts
+  existing paths when require-match is non-nil ([#67], [#125]).
 
 [#4]: https://github.com/raxod502/selectrum/issues/4
 [#12]: https://github.com/raxod502/selectrum/issues/12
@@ -196,6 +198,7 @@ The format is based on [Keep a Changelog].
 [#55]: https://github.com/raxod502/selectrum/issues/55
 [#57]: https://github.com/raxod502/selectrum/pull/57
 [#62]: https://github.com/raxod502/selectrum/pull/62
+[#67]: https://github.com/raxod502/selectrum/issues/67
 [#73]: https://github.com/raxod502/selectrum/pull/73
 [#74]: https://github.com/raxod502/selectrum/pull/74
 [#76]: https://github.com/raxod502/selectrum/pull/76
@@ -214,6 +217,7 @@ The format is based on [Keep a Changelog].
 [#108]: https://github.com/raxod502/selectrum/pull/108
 [#113]: https://github.com/raxod502/selectrum/issues/113
 [#118]: https://github.com/raxod502/selectrum/pull/118
+[#125]: https://github.com/raxod502/selectrum/pull/125
 [#126]: https://github.com/raxod502/selectrum/issues/126
 [#127]: https://github.com/raxod502/selectrum/pull/127
 [raxod502/ctrlf#41]: https://github.com/raxod502/ctrlf/issues/41